fix: delete model

+++ api/interfaces/controllers/user_controller.go
@@ -72,7 +72,7 @@ func (controller *UserController) Save(c echo.Context) (err error) {
 func (controller *UserController) Delete(c echo.Context) (err error) {
 	id, _ := strconv.Atoi(c.Param("id"))
 	user := domain.User{
-		ID: id,
+		ID: uint(id),
 	}
 	err = controller.Interactor.DeleteById(user)
 	if err != nil {
